Prioritize participants in user-mention type-ahead
* Present the usernames of matching users who have previously mapped or
validated in the current project first when offering type-ahead
suggestions for a user-mention in a comment, followed by other users.

* Support an optional `projectId` query parameter to
`user/search/filter` API. The response includes a new `users` array (in
addition to the previous `usernames` for backward compatibility) that is
populated with additional information about each user if a projectId is
provided, including whether that user has "participated" (mapped or
validated) in the project. The `users` array is left empty if no
projectId is given.