Parse new database connection parameters
Problem:

There's no straightforward way to prevent PostgreSQL password
from being displayed in AWS console (our current application host).
This is a shortcoming of AWS Cloudformation and available options
for AWS ECS `ValueFrom` field.

Solution:

Instead of trying to parse the secrets manager entry and pick specific
parameters in the CloudFormation template, we do the parsing in
config.py and pass the whole JSON stored in secrets manager.

This is agnostic of AWS and can be extended to any cloud provider or
server host in the future.