Add IPv6 route to Frontend CloudFront CDN
After enabling IPv6 support on the CDN itself, we need to create DNS
AAAA records to be able to reach the frontend CDN using IPv6. This
change creates the resource records in AWS Route53 - the DNS service for
hotosm.org for both IPv4 and IPv6 routing.

Adds a RecordSetGroup resource along with the RecordSet resource