Add custom exceptions in the list of sentry ignored errors. ----------------------------------------------------
As newly defined custom exceptions are not explicitly handled within the code (because Flask's error handler handles them as they are inherited from Werkzeug's HTTPException), they will be reported as unhandled errors in Sentry. This will result in a significant amount of noise on the Sentry dashboard. To reduce this noise, we need to ignore all these handled errors and exclude them from being reported to Sentry.