Fix #5907: Default changset tags not displayed
This change fixes the problem described below:

Quoting the investigation findings by Hel in issue #5907:

"After investigating the issue with @Aadesh-Baral, the problem appears
to be due to a mismatch between the `TM_DEFAULT_CHANGESET_COMMENT`
environment variable value used in the front end and the back end.
This mismatch causes the front end to expect a different pattern for
the default changeset comment compared to what is being generated by
the backend during project creation.

"It is essential to ensure that the `TM_DEFAULT_CHANGESET_COMMENT`
environment variable is consistent between the front and back end. This
will ensure that the pattern matches the default changeset comment
generated by the backend. For example, if the value of
`TM_DEFAULT_CHANGESET_COMMENT` is set to `#hotosm-project`, then the
project's default changeset comment will be set to `#hotosm-project-XXX`.
But say the frontend environment variable has
`TM_DEFAULT_CHANGESET_COMMENT` as `#hot-project` (or a mismatch), then the
frontend expects the default changeset comment to be in the pattern of
`#hot-project-XXX`. And this mismatch has caused the input field to be
empty, which is the issue."